feat: app/layout.tsx 수정, OAuthAppProvider.tsx 추가
